Fixed the button hover and active
The current implementation causes the layout below the button to jerk up and down when hovering or clicking it. This is because borders affect the element's offset height. My proposal is to use box-shadow to achieve the same effects without any side-effects.

I originally proposed this change for the /r/web_design subreddit but then realized that this is an issue on every subreddit using Naut.